backend/requirements.txt
# --- Core ---
fastapi~=0.115
uvicorn[standard]~=0.30     # ASGI server with reload, websockets, etc.
SQLAlchemy~=2.0
psycopg2-binary~=2.9        # PostgreSQL driver (or use psycopg[binary]~=3.2)

# --- Data validation & settings ---
pydantic~=2.8
pydantic-settings~=2.4      # For env config

# --- Authentication ---
passlib[bcrypt]~=1.7        # Password hashing
python-jose[cryptography]~=3.3  # JWT handling
python-multipart~=0.0.9     # Required for form-based login if you use OAuth2PasswordRequestForm

# --- Dev tools / optional ---
alembic~=1.13               # DB migrations
python-dotenv~=1.0          # Load .env in dev
email-validator~=2.2        # For pydantic EmailStr

